Make algebra more honest and perform node replacement via ana
