+++ core/src/main/scala/geotrellis/store/query/QueryF.scala
@@ -17,17 +17,20 @@
 package geotrellis.store.query
 
 import geotrellis.vector.ProjectedExtent
-
 import io.circe._
 import io.circe.syntax._
 import io.circe.generic.JsonCodec
 import cats.Functor
-import cats.syntax.either._ // scala 2.11 compact
+import cats.syntax.either._
+import cats.data.NonEmptySet
+import cats.instances.string._
 import higherkindness.droste.scheme
 import higherkindness.droste.{Algebra, Coalgebra}
 import higherkindness.droste.syntax.fix._
-
+import higherkindness.droste.syntax.unfix._
 import java.time.ZonedDateTime
+
+import scala.collection.immutable.SortedSet
 
 @JsonCodec sealed trait QueryF[A]
 
@@ -94,6 +97,13 @@ object QueryF {
   }
 
   val coalgebraJson: Coalgebra[QueryF, Json] = Coalgebra(_.foldWith(unfolder))
+
+  def coalgebraWithName(name: String): Coalgebra[QueryF, Query] = coalgebraWithNames(NonEmptySet(name, SortedSet.empty))
+  def coalgebraWithNames(names: NonEmptySet[String]): Coalgebra[QueryF, Query] = Coalgebra {
+    case WithName(_)  => WithName(names.head)
+    case WithNames(_) => WithNames(names.toSortedSet)
+    case e => e.unfix
+  }
 
   def asJson(query: Query): Json  = scheme.cata(algebraJson).apply(query)
   def fromJson(json: Json): Query = scheme.ana(coalgebraJson).apply(json)
