+++ stac-example/src/main/scala/geotrellis/stac/api/SearchFilters.scala
@@ -79,55 +79,25 @@ object SearchFilters {
 
   implicit val searchFilterEncoder: Encoder[SearchFilters] = deriveEncoder
 
-  // dont use the word demo it confuses Eugene
-  // and me
-  //
-  // Question for James and Chris
-  // can it sort htwe return items?
-  // mb it a franklin?>
-  // limit + sort can fix
-  // sort by time, to get the most recent?
-  //
-  // how this should look like?
   import geotrellis.store.query.QueryF._
   def algebra: Algebra[QueryF, SearchFilters] = Algebra {
     // unsupported node
-    case Nothing() => SearchFilters()
-    case All() => SearchFilters()
-    case WithName(name) => SearchFilters(query = Map("layer:ids" -> Map("superset" -> List(name).asJson).asJson).asJsonObject)
-    case WithNames(names) => SearchFilters(query = Map("layer:ids" -> Map("superset" -> names.asJson).asJson).asJsonObject)
-    case At(t, _) => SearchFilters(datetime = TemporalExtent(t.toInstant, None).some)
+    case Nothing()          => SearchFilters()
+    case All()              => SearchFilters()
+    case WithName(name)     => SearchFilters(query = Map("layer:ids" -> Map("superset" -> List(name).asJson).asJson).asJsonObject)
+    case WithNames(names)   => SearchFilters(query = Map("layer:ids" -> Map("superset" -> names.asJson).asJson).asJsonObject)
+    case At(t, _)           => SearchFilters(datetime = TemporalExtent(t.toInstant, None).some)
     case Between(t1, t2, _) => SearchFilters(datetime = TemporalExtent(t1.toInstant, t2.toInstant).some)
-    case Intersects(e) => SearchFilters(intersects = e.reproject(LatLng).extent.toPolygon.some)
+    case Intersects(e)      => SearchFilters(intersects = e.reproject(LatLng).extent.toPolygon.some)
     case Covers(e) =>
       val Extent(xmin, ymin, xmax, ymax) = e.reproject(LatLng).extent
       SearchFilters(bbox = TwoDimBbox(xmin, xmax, ymin, ymax).some)
     // unsupported node
     case Contains(_) => SearchFilters()
     case And(l, r) => l and r
-    // unsupported node // weirdly supported node
-    case Or(l, r) => l.copy(query = r.query)
-  }
-
-  def algebraWithNames(names: List[String]): Algebra[QueryF, SearchFilters] = Algebra {
     // unsupported node
-    case Nothing() => SearchFilters()
-    case All() => SearchFilters()
-    case WithName(_) => SearchFilters(query = Map("layer:ids" -> Map("superset" -> names.take(1).asJson).asJson).asJsonObject)
-    case WithNames(_) => SearchFilters(query = Map("layer:ids" -> Map("superset" -> names.asJson).asJson).asJsonObject)
-    case At(t, _) => SearchFilters(datetime = TemporalExtent(t.toInstant, None).some)
-    case Between(t1, t2, _) => SearchFilters(datetime = TemporalExtent(t1.toInstant, t2.toInstant).some)
-    case Intersects(e) => SearchFilters(intersects = e.reproject(LatLng).extent.toPolygon.some)
-    case Covers(e) =>
-      val Extent(xmin, ymin, xmax, ymax) = e.reproject(LatLng).extent
-      SearchFilters(bbox = TwoDimBbox(xmin, xmax, ymin, ymax).some)
-    // unsupported node
-    case Contains(_) => SearchFilters()
-    case And(l, r) => l and r
-    // unsupported node // weirdly supported node
-    case Or(l, r) => l.copy(query = r.query)
+    case Or(_, _)  => SearchFilters()
   }
 
   def eval(query: Query): SearchFilters = scheme.cata(SearchFilters.algebra).apply(query)
-  def eval(query: Query, names: List[String]): SearchFilters = scheme.cata(SearchFilters.algebraWithNames(names)).apply(query)
 }
